requirements.txt
pandasql
semantic-router[local]
sentence_transformers
streamlit
chromadb
pysqlite3-binary
semantic-router==0.0.20
groq
langchain
langchain-community
langchain-chroma
langchain-groq
langchain-huggingface
python-dotenv
pandas
sentence-transformers